Add oauth server

+++ src/controllers/taskController.js
@@ -4,13 +4,17 @@ const exceptionHandler = require('../helpers/exceptionHandler')
 const taskService = require('../services/taskService')
 
 const router = express.Router()
-const controllerName = 'task'
+const controllerName = 'tasks'
 
 router.get(`/${controllerName}`, async (req, res) => {
   let reponse
   try {
-    reponse = await taskService.getAllTask()
-    return res.status(200).send(reponse).send(console.log("capturing all the data successfully performed"))
+    if (req.headers.oauth) {
+      reponse = await taskService.getAllTask()
+      return res.status(200).send(reponse).send(console.log("capturing all the data successfully performed"))
+    } else {
+      return res.status(403).send("Você Não Tem Permissão Para Acessar Este Servidor (Forbidden – You Don't Have Permission to Access On This Server).")
+    }
   } catch (error) {
     const dateHour = moment().format()
     await exceptionHandler.handleAsync(res, req.header('user-agent'), 'taskGetAll', req.body, reponse, error.message)
@@ -24,13 +28,18 @@ router.get(`/${controllerName}`, async (req, res) => {
 router.get(`/${controllerName}/:id`, async (req, res) => {
   let reponse
   try {
-    const { id } = req.params
-    reponse = await taskService.getById(id)
-    if (reponse.status == true) {
-      return res.status(200).send(reponse.message).send(console.log("capturing all the data successfully performed"))
+    if (req.headers.oauth) {
+      const { id } = req.params
+      reponse = await taskService.getById(id)
+      if (reponse.status == true) {
+        return res.status(200).send(reponse.message).send(console.log("capturing all the data successfully performed"))
+      } else {
+        return res.status(400).send(reponse.message).send(console.log(reponse.message))
+      }
     } else {
-      return res.status(400).send(reponse.message).send(console.log(reponse.message))
+      return res.status(403).send("Você Não Tem Permissão Para Acessar Este Servidor (Forbidden – You Don't Have Permission to Access On This Server).")
     }
+
   } catch (error) {
     const dateHour = moment().format()
     await exceptionHandler.handleAsync(res, req.header('user-agent'), 'taskGetById', req.body, reponse, error.message)
@@ -44,7 +53,8 @@ router.get(`/${controllerName}/:id`, async (req, res) => {
 router.post(`/${controllerName}`, async (req, res) => {
   let reponse
   try {
-    if (req.headers["user-agent"]) {
+
+    if (req.headers.oauth) {
       reponse = await taskService.postNewTask(req.body)
       const dateHour = moment().format()
       if (reponse.statu != false) {
@@ -71,13 +81,17 @@ router.post(`/${controllerName}`, async (req, res) => {
 router.delete(`/${controllerName}/:id`, async (req, res) => {
   let reponse
   try {
-    const { id } = req.params
-    reponse = await taskService.deleteById(id)
-    const dateHour = moment().format()
-    if (reponse == "Ok") {
-      return res.status(200).send("Successfully deleted").send(console.log("Successfully deleted - " + dateHour))
+    if (req.headers.oauth) {
+      const { id } = req.params
+      reponse = await taskService.deleteById(id)
+      const dateHour = moment().format()
+      if (reponse == "Ok") {
+        return res.status(200).send("Successfully deleted").send(console.log("Successfully deleted - " + dateHour))
+      } else {
+        return res.status(400).send("Id not found or error").send(console.log("Id not found or error - " + dateHour))
+      }
     } else {
-      return res.status(400).send("Id not found or error").send(console.log("Id not found or error - " + dateHour))
+      return res.status(403).send("Você Não Tem Permissão Para Acessar Este Servidor (Forbidden – You Don't Have Permission to Access On This Server).")
     }
   } catch (error) {
     const dateHour = moment().format()
@@ -92,6 +106,7 @@ router.delete(`/${controllerName}/:id`, async (req, res) => {
 router.put(`/${controllerName}/:id`, async (req, res) => {
   let reponse
   try {
+    if (req.headers.oauth) {
     const { id } = req.params
     reponse = await taskService.putById(id, req.body)
     const dateHour = moment().format()
@@ -100,6 +115,9 @@ router.put(`/${controllerName}/:id`, async (req, res) => {
     } else {
       return res.status(400).send("Id not found or error").send(console.log("Id not found or error - " + dateHour))
     }
+  } else {
+    return res.status(403).send("Você Não Tem Permissão Para Acessar Este Servidor (Forbidden – You Don't Have Permission to Access On This Server).")
+  }
   } catch (error) {
     const dateHour = moment().format()
     await exceptionHandler.handleAsync(res, req.header('user-agent'), 'taskGetById', req.body, reponse, error.message)
